Move the tests into the right module
