+++ reactivesocket-transport-tcp/src/main/java/io/reactivesocket/transport/tcp/NettyDuplexConnection.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivesocket.netty;
+package io.reactivesocket.transport.tcp;
 
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
