+++ reactivesocket-transport-tcp/src/test/java/io/reactivesocket/transport/tcp/ClientServerTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivesocket.netty.websocket;
+package io.reactivesocket.transport.tcp;
 
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.channel.Channel;
@@ -22,9 +22,6 @@ import io.netty.channel.ChannelPipeline;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.nio.NioEventLoopGroup;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
-import io.netty.handler.codec.http.HttpObjectAggregator;
-import io.netty.handler.codec.http.HttpServerCodec;
-import io.netty.handler.codec.http.websocketx.WebSocketServerProtocolHandler;
 import io.netty.handler.logging.LogLevel;
 import io.netty.handler.logging.LoggingHandler;
 import io.reactivesocket.ConnectionSetupPayload;
@@ -32,8 +29,8 @@ import io.reactivesocket.DefaultReactiveSocket;
 import io.reactivesocket.Payload;
 import io.reactivesocket.ReactiveSocket;
 import io.reactivesocket.RequestHandler;
-import io.reactivesocket.netty.websocket.client.ClientWebSocketDuplexConnection;
-import io.reactivesocket.netty.websocket.server.ReactiveSocketServerHandler;
+import io.reactivesocket.transport.tcp.client.ClientTcpDuplexConnection;
+import io.reactivesocket.transport.tcp.server.ReactiveSocketServerHandler;
 import io.reactivesocket.test.TestUtil;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
@@ -115,24 +112,20 @@ public class ClientServerTest {
                 @Override
                 protected void initChannel(Channel ch) throws Exception {
                     ChannelPipeline pipeline = ch.pipeline();
-                    pipeline.addLast(new HttpServerCodec());
-                    pipeline.addLast(new HttpObjectAggregator(64 * 1024));
-                    pipeline.addLast(new WebSocketServerProtocolHandler("/rs"));
                     pipeline.addLast(serverHandler);
                 }
             });
 
-        serverChannel = b.bind("localhost", 8025).sync().channel();
+        serverChannel = b.bind("localhost", 7878).sync().channel();
 
-        ClientWebSocketDuplexConnection duplexConnection = RxReactiveStreams.toObservable(
-            ClientWebSocketDuplexConnection.create(InetSocketAddress.createUnresolved("localhost", 8025), "/rs", new NioEventLoopGroup())
+        ClientTcpDuplexConnection duplexConnection = RxReactiveStreams.toObservable(
+            ClientTcpDuplexConnection.create(InetSocketAddress.createUnresolved("localhost", 7878), new NioEventLoopGroup())
         ).toBlocking().single();
 
         client = DefaultReactiveSocket
             .fromClientConnection(duplexConnection, ConnectionSetupPayload.create("UTF-8", "UTF-8"), t -> t.printStackTrace());
 
         client.startAndWait();
-
     }
 
     @AfterClass
@@ -202,8 +195,8 @@ public class ClientServerTest {
             .range(1, count)
             .flatMap(i ->
                 RxReactiveStreams
-                    .toObservable(client.requestResponse(
-                        TestUtil.utf8EncodedPayload("hello", "metadata")))
+                    .toObservable(client
+                        .requestResponse(TestUtil.utf8EncodedPayload("hello", "metadata")))
                     .map(payload -> TestUtil.byteToString(payload.getData()))
             )
             .doOnError(Throwable::printStackTrace)
